requirements.txt
numpy
pandas
tqdm
PyYAML
scikit-learn
scipy
matplotlib
jupyter

